requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.0
PyYAML==6.0.1
tkinterdnd2==0.3.0
pytest==7.4.0
